warnings /W4 (windows)

+++ scilab/modules/sound/sci_gateway/c/sci_beep.c
@@ -75,7 +75,7 @@ int sci_Beep(char *fname,unsigned long fname_len)
 void BeepLinuxWindows(void)
 {
 	#ifdef _MSC_VER
-		MessageBeep(-1);
+		MessageBeep((UINT)-1);
 	#else
   /* !!!! This code MUST be rewrited !!! It is not at all the way of playing
 	 sound with Linux */
